Trait moore::arenas::AllocInto

pub trait AllocInto<'t, T>: Alloc<'a, 't, T> where
    T: 't, { }

Allocates values into some arena.

This is merely a marker trait that you should not implement yourself. It is implemented automatically on anything that supports Alloc<'a, 't, T> for any 'a. The allocated values have the lifetime 't.