Module moore::score

This module implements the scoreboard building blocks. Scoreboards are the driving mechanism behind moore. They keep track of the results of each compilation step for every node in the graph. Each node can be accessed in a type safe manner by its ID.

Traits

GenericContext
A context which provides a language-agnostic scoreboard. This is used by the language-specific scoreboards to communicate with the global scoreboard.

NodeMaker
The NodeMaker trait allows for nodes to be generated from an ID.

NodeRef
A reference to a node.

NodeStorage
The NodeStorage trait allows for references to nodes to be stored and retrieved via a unique node ID.

Type Definitions

Result
The result of making a node. Errors that occur while making a node should be reported via a separate channel, e.g. diagnostics, which provide more information to the user.