Trait moore::score::NodeRef

pub trait NodeRef: Eq + Ord + Into<NodeId> + Copy + Debug + Hash {
    fn new(id: NodeId) -> Self;

    fn alloc() -> Self { ... }
}

A reference to a node.

Newtypes around NodeId should implement this trait to offer functionality common to all node references.

Required methods

fn new(id: NodeId) -> Self

Create a new reference from an existing node ID.

Loading content...
Provided methods

fn alloc() -> Self

Allocate a new reference.

Creates a new unique reference. Calls NodeId::alloc() under the hood.

Loading content...