Struct moore::score::ScoreContext

pub struct ScoreContext<'lazy, 'sb: 'lazy, 'ast: 'sb, 'ctx: 'sb> {
    pub sess: &'lazy Session,
    pub sb: &'sb ScoreBoard<'ast, 'ctx>,
    pub vhdl: &'sb ScoreBoard<'ast, 'ctx>,
    pub vhdl_phases: &'lazy LazyPhaseTable<'sb, 'ast, 'ctx>,
    pub svlog: &'sb GlobalContext<'ast>,
}

The global context which holds information about the used scoreboards. All useful operations are defined on this context rather than on the scoreboard directly, to decouple processing and ownership.

Fields
sess: &'lazy Session

The compiler session which carries the options and is used to emit diagnostics.

sb: &'sb ScoreBoard<'ast, 'ctx>

The global scoreboard.

vhdl: &'sb ScoreBoard<'ast, 'ctx>

The VHDL scoreboard.

vhdl_phases: &'lazy LazyPhaseTable<'sb, 'ast, 'ctx>

The VHDL lazy phase table.

svlog: &'sb GlobalContext<'ast>

The SystemVerilog scoreboard.

Methods

impl<'lazy, 'sb, 'ast, 'ctx> ScoreContext<'lazy, 'sb, 'ast, 'ctx>

pub fn vhdl(&'lazy self) -> ScoreContext<'lazy, 'sb, 'ast, 'ctx>

Obtain a reference to the VHDL context.

pub fn add_library(&self, name: Name, asts: &'ast [Ast]) -> LibRef

Add a library to the scoreboard.

pub fn defs(&self, id: ScopeRef) -> Result<&'ctx Defs>

Obtain the definitions in a scope. Calculate them if needed.