Trait moore::score::NodeRef

pub trait NodeRef: Copy + Eq + Ord + Into<NodeId> + Hash + Debug {
    fn new(id: NodeId) -> Self;

    fn alloc() -> Self { ... }
}

A reference to a node.

Newtypes around NodeId should implement this trait to offer functionality common to all node references.

Required Methods

fn new(id: NodeId) -> Self

Create a new reference from an existing node ID.

Provided Methods

fn alloc() -> Self

Allocate a new reference.

Creates a new unique reference. Calls NodeId::alloc() under the hood.