Trait moore::util::HasSpan

pub trait HasSpan {
    fn span(&self) -> Span;

    fn human_span(&self) -> Span { ... }
}

Provides span information for syntax nodes.

Required Methods

fn span(&self) -> Span

Obtain the full span of the input file that this node covers.

Provided Methods

fn human_span(&self) -> Span

Obtain a span which can be used to refer to this node in error messages presented to humans. This will generally be the name for things like entities, processes, and variables. Defaults to return whatever span() returns.